requirements.txt
Flask
gunicorn
requests
sentence-transformers
msal
python-dotenv
openai